backend/requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
fredapi
newsapi-python
python-dotenv
psycopg2-binary
scikit-learn
joblib
shap
lightgbm
nltk
vaderSentiment
google-generativeai
feedparser
requests
beautifulsoup4
lxml
